+++ src/test/java/org/springframework/beans/factory/HelloService.java
@@ -8,8 +8,9 @@ import lombok.Setter;
  * @date: 2024/9/29
  * @project: sbs-small-spring
  */
-public class HelloService {
+public class HelloService implements InitializingBean, DisposableBean {
   @Getter @Setter private String name;
+  @Getter @Setter private String id;
   @Getter @Setter private WorldService worldService;
 
   public HelloService(String name) {
@@ -25,5 +26,15 @@ public class HelloService {
 
   public void sayWorld() {
     worldService.world();
+  }
+
+  @Override
+  public void destroy() throws Exception {
+    System.out.println("执行：HelloService destroy() ");
+  }
+
+  @Override
+  public void afterPropertiesSet() throws Exception {
+    System.out.println("执行：HelloService.afterPropertiesSet");
   }
 }