+++ src/test/java/org/springframework/beans/factory/WorldService.java
@@ -1,12 +1,16 @@
 package org.springframework.beans.factory;
 
+import java.util.HashMap;
+import java.util.Map;
+
 /**
  * @author: DoubleW2w
  * @date: 2024/9/30
  * @project: sbs-small-spring
  */
 public class WorldService {
-  private String id = "world";
+  private String id;
+  private static Map<String, String> hashMap = new HashMap<>();
 
   public WorldService() {}
 
@@ -15,7 +19,19 @@ public class WorldService {
   }
 
   public String world() {
-    System.out.println(id);
     return id;
+  }
+
+  public void initDataMethod() {
+    System.out.println("执行：init-method");
+    this.id = "zhangsan";
+    hashMap.put("10001", "小叮当");
+    hashMap.put("10002", "大叮当");
+    hashMap.put("10003", "笔记本");
+  }
+
+  public void destroyDataMethod() {
+    System.out.println("执行：destroy-method");
+    hashMap.clear();
   }
 }